added damage apply system

+++ Assets/Scripts/ECS Logic/Common/Health/Components/HealthPoints.cs
@@ -4,6 +4,6 @@ namespace ECS_Logic.Common.Health.Components
 {
 	public struct HealthPoints : IComponentData
 	{
-		public float Value;
+		public int Value;
 	}
 }
